pyproject.toml
[tool.poetry]
name = "obs-touchosc"
version = "0.1.0"
description = "OBS Websocket to OSC converter to be used with TouchOSC"
authors = ["Johannes Schriewer <hallo@dunkelstern.de>"]
license = "GPL-2"

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
oscpy = "^0.5.0"
PySide2 = "^5.15.2"
obs-websocket-py = "^0.5.3"

[tool.poetry.dev-dependencies]
pyinstaller = "^4.1"
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
